# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Luckykeeper",
            "url": "https://github.com/luckykeeper/CocoaSyncer",
            "email": "luckykeeper@luckykeeper.site"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/badge": {
            "get": {
                "description": "【浏览器打开，非程序调用】向用户展示 CocoaSyncer 节点卡片（Badge）",
                "tags": [
                    "监控"
                ],
                "summary": "展示 CocoaSyncer 节点卡片，请使用浏览器打开链接，Swagger 无法调用",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/getClusterInfo": {
            "post": {
                "description": "【程序调用】获取节点和集群的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "业务"
                ],
                "summary": "获取集群信息",
                "parameters": [
                    {
                        "description": "API 请求数据，仅传其中的 apiRequest.CocoaSecret 即可完成该接口的鉴权，不需要传其它参数",
                        "name": "apiRequest",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.APIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ServerReturn"
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "【浏览器打开，非程序调用】向用户展示 CocoaSyncer 运行状态",
                "tags": [
                    "监控"
                ],
                "summary": "展示 CocoaSyncer 运行状态，请使用浏览器打开链接，Swagger 无法调用",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/updateClusterInfo": {
            "post": {
                "description": "【程序调用】用于 leader 节点将 follower 节点加入集群，由 leader 节点向 follower 节点推送数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "业务"
                ],
                "summary": "加入集群/更新信息",
                "parameters": [
                    {
                        "description": "API 请求数据，鉴权参数是 apiRequest.CocoaSecret ，节点仅接收 `otherCocoaSyncer`, `cocoaManagedService`, `cloudPlatformInfo`的数据，其中状态码和喂狗数据不接收",
                        "name": "apiRequest",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.APIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ServerReturn"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.APIRequest": {
            "type": "object",
            "properties": {
                "cocoaBasic": {
                    "description": "更新节点信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.CocoaBasic"
                        }
                    ]
                },
                "cocoaSecret": {
                    "description": "节点的请求密钥",
                    "type": "string"
                }
            }
        },
        "model.CloudPlatformInfo": {
            "type": "object",
            "properties": {
                "accessKey": {
                    "description": "DNS API账号的 AccessKey",
                    "type": "string"
                },
                "accessSecret": {
                    "description": "DNS API账号的 AccessSecret",
                    "type": "string"
                },
                "platformProvider": {
                    "description": "云平台服务商",
                    "type": "string"
                }
            }
        },
        "model.CocoaBasic": {
            "type": "object",
            "properties": {
                "apiPort": {
                    "description": "平台相关",
                    "type": "integer"
                },
                "cloudPlatformInfo": {
                    "description": "CocoaSyncer 对接的云平台信息（不在配置文件里面配置）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CloudPlatformInfo"
                    }
                },
                "cocoaManagedService": {
                    "description": "CocoaSyncer 管理的服务（全部）（不在配置文件里面配置）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CocoaManagedService"
                    }
                },
                "cocoaSecret": {
                    "description": "CocoaSyncer 通信密钥，各节点可不同",
                    "type": "string"
                },
                "configImported": {
                    "type": "boolean"
                },
                "createdAt": {
                    "description": "表自身的属性",
                    "type": "string"
                },
                "dataBaseType": {
                    "description": "数据库",
                    "type": "string"
                },
                "dsn": {
                    "description": "dsn ，数据库连接信息",
                    "type": "string"
                },
                "nodeAddress": {
                    "description": "CocoaSyncer 节点的访问地址",
                    "type": "string"
                },
                "nodeCode": {
                    "description": "CocoaSyncer 节点代号（WEB 不展示，建议填写地区）",
                    "type": "string"
                },
                "nodeName": {
                    "description": "CocoaSyncer 节点友好名称（WEB 展示及记忆用）",
                    "type": "string"
                },
                "otherCocoaSyncer": {
                    "description": "CocoaSyncer 互联的节点【包含本机】（不在配置文件里面配置）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NodeInfo"
                    }
                },
                "reverseProxy": {
                    "description": "反向代理(true=使用)",
                    "type": "boolean"
                },
                "timeout": {
                    "description": "超时时间",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workMode": {
                    "description": "CocoaSyncer 工作模式（leader|follower）",
                    "type": "string"
                }
            }
        },
        "model.CocoaManagedService": {
            "type": "object",
            "properties": {
                "domain": {
                    "description": "服务域名名称(e.g. luckykeeper.site)",
                    "type": "string"
                },
                "platformProvider": {
                    "description": "云平台服务商",
                    "type": "string"
                },
                "record": {
                    "description": "主机名称(e.g. www)",
                    "type": "string"
                },
                "serviceName": {
                    "description": "服务友好名称",
                    "type": "string"
                }
            }
        },
        "model.NodeInfo": {
            "type": "object",
            "properties": {
                "ISPs": {
                    "description": "域名解析线路 (default|CT|CU|CMCC|CNEdu|OverSea|None)[默认|中国电信|中国联通|中国移动|中国教育网|境外|不解析]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "cocoaManagedService": {
                    "description": "节点管理的服务",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CocoaManagedService"
                    }
                },
                "cocoaSecret": {
                    "description": "CocoaSyncer 通信密钥，各节点可不同",
                    "type": "string"
                },
                "networkType": {
                    "description": "节点网络类型（IPV4|IPV6|Both）[仅使用IPV4|仅使用IPV6|IPV4\u0026IPV6双栈]",
                    "type": "string"
                },
                "nodeAddress": {
                    "description": "CocoaSyncer 节点的访问地址",
                    "type": "string"
                },
                "nodeCode": {
                    "description": "CocoaSyncer 节点代号（WEB 不展示，建议填写地区）",
                    "type": "string"
                },
                "nodeName": {
                    "description": "CocoaSyncer 节点友好名称（WEB 展示及记忆用）",
                    "type": "string"
                },
                "statusCode": {
                    "description": "节点状态码，200 正常，504 失联",
                    "type": "integer"
                },
                "watchDog": {
                    "description": "看门狗，判断master是否存活",
                    "type": "string"
                },
                "workMode": {
                    "description": "CocoaSyncer 工作模式（leader|follower）",
                    "type": "string"
                }
            }
        },
        "model.ServerReturn": {
            "type": "object",
            "properties": {
                "cocoaBasic": {
                    "description": "业务数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.CocoaBasic"
                        }
                    ]
                },
                "statusCode": {
                    "description": "业务码 （操作成功200，Token错误401）",
                    "type": "integer"
                },
                "statusString": {
                    "description": "业务码的文字说明",
                    "type": "string"
                }
            }
        }
    }
}